requirements.txt
django===3.0.4
djangorestframework===3.7.3
django-cors-headers===2.1.0